simplify examples structure (#1247)
* simplify examples structure

* update changelog

* fix imports

* rename example

* rename scripts

* changelog
