+++ pl_examples/__init__.py
@@ -140,7 +140,7 @@ Hyperparameter search on a SLURM HPC cluster
 
 """
 
-from .basic_examples.lightning_module_template import LightningTemplateModel
+from pl_examples.models.lightning_template import LightningTemplateModel
 
 __all__ = [
     'LightningTemplateModel'
