+++ pl_examples/basic_examples/submit_ddp2_job.sh
