+++ pl_examples/domain_templates/generative_adversarial_net.py
@@ -1,6 +1,6 @@
 """
 To run this template just do:
-python gan.py
+python generative_adversarial_net.py
 
 After a few epochs, launch TensorBoard to see the images being generated at every batch:
