+++ pl_examples/domain_templates/imagenet.py
