+++ pl_examples/models/lightning_template.py
